virulence function as module file
